pyproject.toml
[tool.poetry]
name = "vapidemo"
version = "0.1.0"
description = "A demo of VAPI, Datatrails, OpenAI and the Conserver"
authors = ["Thomas Howe <ghostofbasho@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
vcon = "^0.3.5"
streamlit-javascript = "^0.1.5"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
requests = "^2.32.3"
aiofiles = "^24.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
